transaction mutation + refactoring

+++ lib/bloc/login_bloc.dart
@@ -1,5 +1,5 @@
 import 'package:bloc/bloc.dart';
-import 'package:fa_bank/api/api_repository.dart';
+import 'package:fa_bank/api/repository.dart';
 import 'package:fa_bank/injector/injector.dart';
 import 'package:fa_bank/podo/login/login_body.dart';
 import 'package:fa_bank/podo/token/token.dart';
@@ -54,7 +54,7 @@ class LoginBloc extends Bloc<LoginEvent, LoginState> {
     await sharedPreferencesManager.putBool(SharedPreferencesManager.keyIsLogin, true);
     await sharedPreferencesManager.putInt(SharedPreferencesManager.keyAuthMSecs, DateTime.now().millisecondsSinceEpoch);
     await sharedPreferencesManager.putString(SharedPreferencesManager.keyUsername, loginBody.username);
-    await sharedPreferencesManager.putInt(SharedPreferencesManager.keyUid, 10527075); // 10527024
+    await sharedPreferencesManager.putDouble(SharedPreferencesManager.keyUid, 10527075); // 10527024
     yield LoginSuccess();
   }
